finalmente funziona

+++ src/tile.cpp
@@ -1,9 +1,9 @@
 #include "tile.h"
 
-int surject(double x, int precision){
+int surject(double x, float precision){
     return (int)(pow(10, precision) * x);
 }
 
-double inverse_surject(int x, int precision){
+double inverse_surject(int x, float precision){
     return (double)x / pow(10, precision);
 }
